feat: Move jobs to a separate endpoint
